+++ src/scene/loop/AnimationFrameLoop.js
@@ -4,19 +4,19 @@
 
 export class AnimationFrameLoop {
   /**
-   * @param {AnimationFrameLoopCallback} callback
    * @param {object} [opts]
-   * @param {Window} [opts.animationFrameHandler]
+   * @param {AnimationFrameLoopCallback} [opts.callback]
+   * @param {Window} [opts.animationFrameSource]
    */
-  constructor(callback, opts = undefined) {
-    const { animationFrameHandler = window } = opts || {};
+  constructor(opts = {}) {
+    const { callback = () => {}, animationFrameSource = window } = opts;
 
     /** @type {ReturnType<requestAnimationFrame>} */
     this.handle = 0;
     this.detail = new AnimationFrameDetail();
 
     /** @protected */
-    this.animationFrameHandler = animationFrameHandler;
+    this.animationFrameSource = animationFrameSource;
 
     /** @protected */
     this.callback = callback;
@@ -30,8 +30,9 @@ export class AnimationFrameLoop {
     return this.handle !== 0;
   }
 
+  /** @protected */
   next(now = performance.now()) {
-    this.handle = this.animationFrameHandler.requestAnimationFrame(this.next);
+    this.handle = this.animationFrameSource.requestAnimationFrame(this.next);
     let d = this.detail;
     d.prevTime = d.currentTime;
     d.currentTime = now;
@@ -39,13 +40,19 @@ export class AnimationFrameLoop {
     this.callback(this);
   }
 
-  start() {
-    this.handle = this.animationFrameHandler.requestAnimationFrame(this.next);
+  /**
+   * @param {AnimationFrameLoopCallback} [callback]
+   */
+  start(callback = undefined) {
+    if (typeof callback !== 'undefined') {
+      this.callback = callback;
+    }
+    this.handle = this.animationFrameSource.requestAnimationFrame(this.next);
     return this;
   }
 
   cancel() {
-    this.animationFrameHandler.cancelAnimationFrame(this.handle);
+    this.animationFrameSource.cancelAnimationFrame(this.handle);
     return this;
   }
 }